+++ lib/src/circle_button.dart
@@ -1,17 +1,48 @@
 import 'package:assorted_layout_widgets/assorted_layout_widgets.dart';
 import "package:flutter/material.dart";
 
+typedef CircleButtonBuilder = Widget Function({
+  required bool isHover, // True when the mouse is over the button.
+  required bool isPressed, // True when the button is tapped-down.
+  required Widget child, // The widget that's displayed when there is no builder.
+});
+
 /// Circular button, similar to IconButton, but with the following differences:
 ///
-/// 1) A [clickAreaMargin] can be set, defining a rectangle around the circle, and all that
-/// rectangle is clickable, so that it's easier to click the button.
+/// - The [icon] will be displayed inside the button.
 ///
-/// 2) The [tapColor] is shown in the tap-down, faster than the GestureDetector.
+/// - A [clickAreaMargin] can be set, defining a rectangle around the circle, and all
+///   that rectangle is clickable, so that it's easier to click the button.
 ///
-/// 3) if [onTap] is null, still the tapColor will be applied. This allows the button to be used
-/// inside of other widgets that detect gestures.
+/// - The [backgroundColor]: The background color of the button.
 ///
-/// 4) Pass [debugShowClickableArea] true to see the clickable area.
+/// - The [tapColor] is shown when the button is tapped-down.
+///
+/// - The [hoverColor] is shown when the mouse is over the button.
+///
+/// - [elevation]: The elevation of the button.
+///
+/// - The [onTap] is the callback called when the button is tapped. If it's null,
+///   the tapColor will still be applied. This allows the button to be used inside of
+///   other widgets that detect gestures.
+///
+/// - [size]: The size of the button.
+///
+/// - [iconPadding]: The padding around the icon inside the button.
+///
+/// - [border]: The border of the button.
+///
+/// - [cursor]: The mouse cursor to be shown when hovering over the button.
+///
+/// - [clickAreaMargin]: The margin around the circle that is also clickable.
+///
+/// - [colorOfClickableArea] is the color of the clickable area, which is only shown
+///   when [debugShowClickableArea] is set to true, for debugging purposes. Otherwise,
+///   it's transparent.
+///
+/// - [builder]: An optional custom builder function to modify the button's child widget.
+///   This can be used to animate the button when the button is tapped, or when the
+///   mouse is over it.
 ///
 class CircleButton extends StatefulWidget {
   //
@@ -31,6 +62,47 @@ class CircleButton extends StatefulWidget {
   final double? elevation;
   final MouseCursor? cursor;
 
+  /// The builder can be used to animate the button when the button is tapped,
+  /// or when the mouse is over it.
+  final CircleButtonBuilder? builder;
+
+  /// Circular button, similar to IconButton, but with the following differences:
+  ///
+  /// - The [icon] will be displayed inside the button.
+  ///
+  /// - A [clickAreaMargin] can be set, defining a rectangle around the circle, and all
+  ///   that rectangle is clickable, so that it's easier to click the button.
+  ///
+  /// - The [backgroundColor]: The background color of the button.
+  ///
+  /// - The [tapColor] is shown when the button is tapped-down.
+  ///
+  /// - The [hoverColor] is shown when the mouse is over the button.
+  ///
+  /// - [elevation]: The elevation of the button.
+  ///
+  /// - The [onTap] is the callback called when the button is tapped. If it's null,
+  ///   the tapColor will still be applied. This allows the button to be used inside of
+  ///   other widgets that detect gestures.
+  ///
+  /// - [size]: The size of the button.
+  ///
+  /// - [iconPadding]: The padding around the icon inside the button.
+  ///
+  /// - [border]: The border of the button.
+  ///
+  /// - [cursor]: The mouse cursor to be shown when hovering over the button.
+  ///
+  /// - [clickAreaMargin]: The margin around the circle that is also clickable.
+  ///
+  /// - [colorOfClickableArea] is the color of the clickable area, which is only shown
+  ///   when [debugShowClickableArea] is set to true, for debugging purposes. Otherwise,
+  ///   it's transparent.
+  ///
+  /// - [builder]: An optional custom builder function to modify the button's child widget.
+  ///   This can be used to animate the button when the button is tapped, or when the
+  ///   mouse is over it.
+  ///
   const CircleButton({
     Key? key,
     required this.icon,
@@ -44,6 +116,7 @@ class CircleButton extends StatefulWidget {
     this.iconPadding,
     this.elevation,
     this.cursor = SystemMouseCursors.click,
+    this.builder,
     bool debugShowClickableArea = false,
   })  : colorOfClickableArea =
             debugShowClickableArea ? const Color(0xBBFF0000) : const Color(0x00000000),
@@ -54,6 +127,33 @@ class CircleButton extends StatefulWidget {
 }
 
 class _CircleButtonState extends State<CircleButton> {
+  //
+  Widget get icon => widget.icon;
+
+  EdgeInsetsGeometry? get clickAreaMargin => widget.clickAreaMargin;
+
+  EdgeInsetsGeometry? get iconPadding => widget.iconPadding;
+
+  Border? get border => widget.border;
+
+  Color? get tapColor => widget.tapColor;
+
+  Color? get hoverColor => widget.hoverColor;
+
+  Color get colorOfClickableArea => widget.colorOfClickableArea;
+
+  Color get backgroundColor => widget.backgroundColor;
+
+  VoidCallback? get onTap => widget.onTap;
+
+  double get size => widget.size;
+
+  double? get elevation => widget.elevation;
+
+  MouseCursor? get cursor => widget.cursor;
+
+  CircleButtonBuilder? get builder => widget.builder;
+
   bool isHover = false;
 
   void _onEnter(_) {
@@ -74,51 +174,58 @@ class _CircleButtonState extends State<CircleButton> {
     return MouseRegion(
       onEnter: _onEnter,
       onExit: _onExit,
-      cursor: widget.cursor ?? MouseCursor.defer,
+      cursor: cursor ?? MouseCursor.defer,
       child: Button(
-        onTap: widget.onTap,
+        onTap: onTap,
         minVisualTapDuration: CircleButton.minVisualTapDuration,
         builder: ({required bool isPressed}) {
-          var child = widget.icon;
-
-          if (widget.iconPadding != null)
-            child = Padding(
-              padding: widget.iconPadding!,
-              child: widget.icon,
-            );
-
           return Container(
-            color: widget.colorOfClickableArea,
-            padding: widget.clickAreaMargin,
-            child: Container(
-                height: widget.size,
-                width: widget.size,
-                decoration: BoxDecoration(
-                  border: widget.border,
-                  color: isPressed
-                      ? widget.tapColor
-                      : (isHover
-                          ? (widget.hoverColor ?? widget.backgroundColor)
-                          : widget.backgroundColor),
-                  shape: BoxShape.circle,
-                  boxShadow: (isPressed || widget.elevation == null || widget.elevation == 0.0)
-                      ? null
-                      : [
-                          BoxShadow(
-                            color: const Color(0x60000000),
-                            offset: Offset(widget.elevation! / 2, widget.elevation!),
-                            blurRadius: widget.elevation!,
-                          ),
-                          BoxShadow(
-                            color: const Color(0xFFFFFFFF),
-                            blurRadius: widget.elevation!,
-                          ),
-                        ],
-                ),
-                child: child),
-          );
+              color: colorOfClickableArea,
+              padding: clickAreaMargin,
+              child: (builder == null)
+                  ? _content(isPressed)
+                  : builder!(
+                      isHover: isHover,
+                      isPressed: isPressed,
+                      child: _content(isPressed)));
         },
       ),
     );
+  }
+
+  Container _content(bool isPressed) {
+    return Container(
+      height: size,
+      width: size,
+      decoration: _decoration(isPressed),
+      child: icon,
+    );
+  }
+
+  BoxDecoration _decoration(bool isPressed) {
+    return BoxDecoration(
+      border: border,
+      color: isPressed
+          ? tapColor
+          : (isHover ? (hoverColor ?? backgroundColor) : backgroundColor),
+      shape: BoxShape.circle,
+      boxShadow: _boxShadow(isPressed),
+    );
+  }
+
+  List<BoxShadow>? _boxShadow(bool isPressed) {
+    return (isPressed || elevation == null || elevation == 0.0)
+        ? null
+        : [
+            BoxShadow(
+              color: const Color(0x60000000),
+              offset: Offset(elevation! / 2, elevation!),
+              blurRadius: elevation!,
+            ),
+            BoxShadow(
+              color: const Color(0xFFFFFFFF),
+              blurRadius: elevation!,
+            ),
+          ];
   }
 }